+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Change Log
 5.6.2
 =====
 * Minor changes to the foursight_core/scripts/local_check_execution.py script.
+  - Fix for --stage prod where we need to set os.environ["chalice_stage"]
+    so that the bucket is crected correctly (e.g. foursight-prod-mastertest).
+  - Make it guess the AWS credentials name from AWS_PROFILE if no --env given.
 
 
 5.6.1